Extrat WebViewRenderer
